Fixing _envsetup
It was crashing at environ = NULL all the time.

First round of malloc. Bucket can't be unsigned long,
it must be signed and ssize_t produces an error.

Migrating to an sbrk with uint64_t arg. Native must be
moved too.

+++ lib/ap/plan9/malloc.c
@@ -36,9 +36,9 @@ struct Arena
 };
 static Arena arena;
 
-#define datoff		((size_t)((Bucket*)0)->data)
+#define datoff		((int64_t)((Bucket*)0)->data)
 
-extern	void	*sbrk(uint32_t);
+extern	void	*sbrk(uint64_t);
 
 void*
 malloc(size_t size)
